Added user to garment, changed all Longs to Integers, started updating controllers to handle request DTOs
TODO update the rest of the controllers so they don't return the entities, but instead return the corresponding request DTO. Also TODO, figure out the lazy loading issue

+++ src/main/java/com/fitcheckme/FitCheckMe/DTOs/Outfit/OutfitUpdateRequestDTO.java
@@ -1,7 +1,7 @@
 package com.fitcheckme.FitCheckMe.DTOs.Outfit;
 
 public record OutfitUpdateRequestDTO (
-	Long outfitId,
+	Integer outfitId,
 	String outfitName,
 	String outfitDesc
 	//TODO Think out how to update the garments
