+++ src/main/java/com/fitcheckme/FitCheckMe/repositories/OutfitRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fitcheckme.FitCheckMe.models.Outfit;
 
-public interface OutfitRepository extends JpaRepository<Outfit, Long> {}
+public interface OutfitRepository extends JpaRepository<Outfit, Integer> {}
